Fixed issues with x_open_display and added documentation.
Previously, it was possible for a Display returned by x_open_display to hold a null-pointer which could potentially result in a segfault when used with other xlib functions. This issue has been fixed by allowing the returned Result to hold a value of None.

Additionally the error returned by x_open_display should be more meaningful now.